Improved Movement
The ship now moves to all of the enemies after clicking on the Confirm Code button, but there is a bug. The ship shoots only at the leftmost and rightmost enemies, and it shoots 3 times at the leftmost enemy (instead of directing the 2nd and 3rd shot at the 2nd and 3rd enemy).